HourRecyclerView now shows correct list - Fixed layout bugs

+++ app/src/main/java/com/example/neoweather/model/weather/NeoWeatherModel.kt
@@ -9,7 +9,7 @@ data class NeoWeatherModel(
 )
 
 data class CurrentWeather(
-    val temperature: Float,
+    val temperature: Double,
     @Json(name = "windspeed") val windSpeed: Double,
     @Json(name = "winddirection") val windDirection: Int,
     @Json(name = "weathercode") val weatherCode: Int,
